+++ public/data/timetabledata.json
@@ -1,0 +1,65 @@
+ [
+  {
+    "day": "MON",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3", "Dish3", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "TUE",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "WED",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "THURS",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "FRI",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "SAT",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  },
+  {
+    "day": "SUN",
+    "dishes": [
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2"],
+      ["Dish1", "Dish2", "Dish3"],
+      ["Dish1", "Dish2", "Dish3"]
+    ]
+  }
+]
