+++ src/Components/index.jsx
@@ -1,4 +1,5 @@
 import Button from "./Button/Button";
 import Navbar from "./Navbar/Navbar";
 import Input from "./Input/Input";
-export { Button, Navbar, Input };
+import Contents from "./Timetable/Contents";
+export { Button, Navbar, Input, Contents };
