+++ src/Pages/TimeTable/TimeTable.jsx
@@ -1,4 +1,4 @@
-import Contents from "./Contents";
+import { Contents } from "../../Components";
 import style from "./TimeTable.module.scss";
 
 const TimeTable = () => {
@@ -7,11 +7,11 @@ const TimeTable = () => {
       <h2>Live TimeTable</h2>
       <table>
         <tr>
-          <th></th>
+          <th className={style.radius1}></th>
           <th>Breakfast</th>
           <th>Lunch</th>
           <th>Snacks</th>
-          <th>Dinner</th>
+          <th className={style.radius2}>Dinner</th>
         </tr>
         <Contents />
       </table>
